pyproject.toml
[tool.poetry]
name = "research-agent"
version = "1.0.0"
description = "Autonomous AI research tracking agent"
authors = ["Lee Gonzales <lee@catalystai.services>"]
readme = "README.md"
packages = [{include = "research_agent"}]

[tool.poetry.dependencies]
python = "^3.10"
anthropic = "^0.39.0"
click = "^8.1.0"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"
feedparser = "^6.0.10"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
arxiv = "^2.1.0"
lxml = "^5.1.0"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
ruff = "^0.0.284"

[tool.poetry.scripts]
research-agent = "research_agent.cli.main:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
